Add primops and primtypes for nf data structures
Add a primitive type Struct#, with operations structNew#, structAppend#
and structResize#. The type is mutable and all operations happen
in the IO monad (or rather, they thread through a State# RealWorld).
The root pointer was removed and is now stored in the wrapper
Struct (which is a normal Haskell ADT).