Rename Struct to Compact
In the mailing list it was decided that the name Struct
is not descriptive enough, so we went with Compact
(StgCompactNFData, COMPACT_NFDATA in the rts because
StgCompact alone was confusing to me, and CNF.[ch]
because Compact.[ch] already existed with an unrelated
purpose). All API has been changed to reflect this,
although variable and argument names are still s/str.