CNF: use a hash table for sharing instead of forwarding pointers
Unfortunately, the forwarding pointers trick is not thread safe,
and this lack of thread safety is not acceptable (as it affects
the ability of other threads to operate on data that is apparently
unrelated because it's only read, as far as the programmer sees it).
So instead we use a hash table that is local to the append process.