Add noshare variants of Compact operations
Add compactAppendNoShare and compactNewNoShare, which as their
name imply do not preserve sharing (so they might copy the same
object multiple times, and are unsuitable when copying a
data structure which has a loop). Their advantage is that they
do not need to track which objects they already copied, and
therefore they are faster.