CNF: rework compact resizing
Instead of storing compacts as a single consecutive block
of memory, now compacts are a linked list of block groups.
Compacts themselves are linked through the bdescr link
fields, while the blocks within the compacts are linked
through a StgCompactNFDataBlock structure which is stored
at the beginning of the block (which is also used to
retrieve the real compact closure given an arbitrary pointer
inside the block, which is necessary at GC time).