CNF: first stab at incremental Compact allocation
New primop, compactAppendOne, which does one level of evacing
in the compact (plus any indirection adjusting if necessary),
assuming all pointers are already inside it (and in a debug build
it would complain loudly otherwise).
New type class, Compactable, which does manual unfolding of objects
and appends one by one using compactAppendOne (but also checks at
each step if anything was already evaluated and is in a compact) -
this alone avoids the repeated deepseq problem, but it is probably
slower in that the calls are recursive instead of using the GC
algorithm. And that's why it can be sidestepped by using an
instance that calls the old compactAppend on a deepseq-ed object.
New monad, CompactM, which runs the code so that all values are
appended in the Compact as soon as they are generated.

No new tests, yet.