Add a striped allocator
The striped allocator divides the address space in one private
range and a number of chunks. The idea is that each process/machine
is assigned one chunk, and can allocate into other processes
into its own chunk at predictable addresses. This will be used
for O(1) importing of CNF.

Note that this only addresses the megablock side, and does not yet
permit allocation at a given address.