BlockAlloc: implement allocGroupAt()
Use a bitmask to represent the state of a given mblock (invalid,
free, alloced, or part of a larger group) at an arbitrary address.
With that, we know if we can read the block descriptors freely
for a given block, and then allocate at the given address simply
by removing from the free list.