CNF: evacuate static structures in the Compact
Treat static structures the same as non static objects,
and copy them inside the Compact buffer.

Note that this would result in objects with a STATIC closure
type that do not have valid static lists, because they do
not live in the static data. We rely on Evac doing the
HEAP_ALLOCED() check instead of checking the info type,
and then evacuate_compact() to never check the info type.
Sanity/Profiling treat Compacts specially and therefore
should not be a problem.