CNF: fix bugs related to autoexpanding
Turns out that if you don't test your code, it will not work.
So add a test, and make it work.

There were actually a number of bugs related to statistics
with chained blocks, which were not uncovered because tests
do not compile with -debug.

And at the same time consolidate the code a little.