Md5: fix inclusion of stg headers
We cannot include stg/Types.h alone, we need ghcconfig.h first,
but at that point it is safer to use HsFFI.h, which is designed
to have everything we need without the Stg.h baggage.
This fixes the configure script in base which was picking up
the size of StgMD5Context as 0 (obviously bad).