BlockAlloc: fix bugs in allocGroupAt
We were not initializing the size of the slop correctly in case
the first At allocation was in the middle of a mblock (which
never triggered in practice because they would happen in the same
order as a regular non At allocation would give out addresses),
we were not initializing the allocated group for the second
At allocation, and we were computing the wrong free list for
the slops.

Also the MBlock iteration code was completely busted.