Merge Compacts into GHC
Summary:
Compacts are a new data structure that was illustrated at ICFP 2015.
It is an immutable NF data structure with no outgoing pointers,
which can be serialized and GCed quickly.

Simon Marlow had a preliminary review of this code during the
conference, but the code was not in a fully reviewable state
at the time. This is the cleaned up version.

Test Plan:
validate
(new test cases included)

Reviewers: ezyang, simonmar, hvr, bgamari, austin

Subscribers: vikraman, Yuras, RyanGlScott, qnikst, mboes, facundominguez, rrnewton, thomie, erikd

Differential Revision: https://phabricator.haskell.org/D1264

GHC Trac Issues: #11493